Allow custom endpoints when using wasabi
When using wasabi, different regions expect different endpoints. Currently, setting useWasabi to true results in the use of the s3.wasabisys.com endpoint which is meant for the Wasabi US East 1 (N. Virginia) region.  Using the wrong endpoint for a region gives an S3 auth error with a message of expecting a different region. A simple null check on the endpoint var allows the selection of different regions with the correct endpoint.